+++ src/main/core/projects/settings/share-project-settings-to-config.test.ts
@@ -1,10 +1,10 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest';
-import { computeProjectSettingsOverrideState } from './project-settings-override-state';
+import { computeProjectSettingsOverrideState } from './sharing/project-settings-override-state';
 import {
   getProjectSettingsWriteTargets,
   resolveAllProjectSettingsTargets,
-} from './project-settings-target-resolver';
-import { shareProjectSettingsToConfig } from './share-project-settings-to-config';
+} from './sharing/project-settings-target-resolver';
+import { shareProjectSettingsToConfig } from './sharing/share-project-settings-to-config';
 
 const mocks = vi.hoisted(() => ({
   select: vi.fn(),
